backend/pyproject.toml
[tool.poetry]
name = "individual-task-buffed"
version = "0.1.0"
description = ""
authors = ["uandriyko <uandriyko@github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.104.1"
uvicorn = "^0.24.0.post1"
motor = "^3.3.2"
environs = "^9.5.0"
bcrypt = "^4.0.1"
pyjwt = "^2.8.0"
apscheduler = "^3.10.4"
httpx = "^0.25.1"
pytest = "^7.4.3"
trio = "^0.23.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
